requirements.txt
gradio
tensorflow-cpu
librosa
numpy
scikit-learn
matplotlib
soundfile
reportlab